+++ deploy/byron_staging.js
@@ -10,4 +10,7 @@ process.env.GRAPHQL_PORT = '443'
 process.env.GRAPHQL_WEBSOCKET_PROTOCOL = 'ws'
 process.env.GRAPHQL_WEBSOCKET_HOST = 'cardano-graphql-byron.daedalus-operations.com'
 
+const branch = process.env.BRANCH || 'develop'
+process.env.BUCKET = `byron-staging-${branch}-explorer`
+
 deploy()
